Refactor tests to prevent static TLS problem
